+++ src/settings/index.js
@@ -3,7 +3,6 @@ import { FormattedMessage } from 'react-intl';
 import { Settings } from '@folio/stripes/smart-components';
 
 import GeneralSettings from './GeneralSettings';
-import OrgRoleSettings from './OrgRoleSettings';
 
 export default class ErmSettings extends React.Component {
   pages = [
@@ -11,11 +10,6 @@ export default class ErmSettings extends React.Component {
       route: 'general',
       label: <FormattedMessage id="ui-agreements.settings.general" />,
       component: GeneralSettings,
-    },
-    {
-      route: 'org-roles',
-      label: <FormattedMessage id="ui-agreements.settings.orgRoles" />,
-      component: OrgRoleSettings,
     },
   ];